Renamed vm-ui-plugin to vm-service and added metadata
