--- samples/react-extension/.dockerignore
@@ -1,1 +1,0 @@
-client/node_modules
