--- samples/react-extension/client/.editorconfig
@@ -1,6 +1,0 @@
-[*]
-root = true
-end_of_line = lf
-
-indent_size = 2
-indent_style = space
