--- samples/react-extension/client/.env
@@ -1,2 +1,0 @@
-BUILD_PATH=dist
-PUBLIC_URL=.
