--- samples/react-extension/client/.prettierrc.json
@@ -1,4 +1,0 @@
-{
-	"singleQuote": true,
-	"trailingComma": "all"
-}
