--- samples/react-extension/client/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@docker/react-extension",
-  "version": "0.2.0",
-  "private": true,
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build"
-  },
-  "dependencies": {
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@docker/docker-mui-theme": "^0.0.4",
-    "@docker/extension-api-client": "^0.2.2",
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.4.1",
-    "@mui/material": "^5.4.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "typescript": "^4.1.2"
-  },
-  "devDependencies": {
-    "@docker/extension-api-client-types": "^0.2.2"
-  }
-}
