--- samples/react-extension/client/public/index.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-  </head>
-  <body class="dockerDesktopTheme">
-    <div id="root"></div>
-  </body>
-</html>
