--- samples/react-extension/metadata.json
@@ -1,10 +1,0 @@
-{
-  "icon": "docker.svg",
-  "ui": {
-    "dashboard-tab": {
-      "title": "UI Extension",
-      "root": "/ui",
-      "src": "index.html"
-    }
-  }
-}
