+++ samples/vm-service/docker-compose.yaml
