+++ samples/vm-service/ui/src/index.html
@@ -5,9 +5,9 @@
   </head>
 
   <body class="dockerDesktopTheme">
-    <h1>Volumes plugin sample</h1>
+    <h1>Volumes extension sample</h1>
 
-    <p onclick="window.ddClient.desktopUI.toast.error('plugin custom error');">
+    <p onclick="window.ddClient.desktopUI.toast.error('custom error');">
       click to fake an error
     </p>
