+++ samples/vm-service/ui/src/script.js
