+++ samples/vm-service/vm/internal/version/version.go
